Windows: detach for real, liveness that is conclusive, identity by file index
A second model reviewed the Windows build (five findings, all taken):

  - detach left the child attached and said nothing; it now asks for a
    new process group and no console (DETACHED_PROCESS), which is what
    setsid buys on unix;
  - Alive read the exit code and treated 259 as running, and a process
    can exit with 259; it now waits on the handle for zero milliseconds,
    which is the one conclusive test, and refuses a pid the kernel
    cannot name rather than truncating it;
  - the file identity on Windows was mtime alone; it is now the volume
    serial, the file index and the creation time, which is what the
    filesystem calls a file as opposed to what it is named;
  - ctimeOf existed for darwin and linux only, so every other unix target
    failed to build internal/paths; the remaining targets carry a zero
    and the linux timespec arithmetic is widened first, since the fields
    are int32 on some architectures (linux/386 vets now).

+++ internal/paths/fileidentity_linux.go
@@ -4,4 +4,7 @@ package paths
 
 import "syscall"
 
-func ctimeOf(st *syscall.Stat_t) int64 { return st.Ctim.Sec*1e9 + st.Ctim.Nsec }
+func ctimeOf(st *syscall.Stat_t) int64 {
+	// Widened first: the fields are int32 on some architectures.
+	return int64(st.Ctim.Sec)*1e9 + int64(st.Ctim.Nsec)
+}
